Increase stdout buffer read size maximum to 30MB
Sometimes the JSON output from `npm audit` can be quite large,
especially for project with a lot of dependencies, so let's increase
the max size we're willing to read from stdout before truncating.
Truncated JSON will almost always result in a parser error when we
call `JSON.parse`, so this change increases the usefulness of the
output of this script for larger projects.